3_dockerize/pulsar-testing/django/requirements_django.txt
Django==1.11.29
django-appconf==1.0.2
django-bower==5.1.0
django-compressor==2.1
django-el-pagination==2.1.1
django-modelcluster==2.0
django-taggit==0.18.3
pytz==2016.6.1
six==1.10.0
Unidecode==0.4.19
wheel==0.24.0
pulsar-client==2.3.2
